test(renderer): split the #2441 camera tests on their subject seam
`camera-degenerate-pose.test.ts` had grown to cover two units with two
different failure stories that only share the #2441 origin:

  - a degenerate pose must not turn the view-projection matrix into NaN
    (`MathUtils.lookAt`, near/far, and the `orthoSize`/`fov`/`aspect`
    inputs that feed the projection alongside the pose) — the matrix
    derivation in `camera.ts` / `math.ts`;
  - a navigation gesture must not spread a malformed pose it read back
    out (pan, orbit on both pivots, zoom, first-person walk,
    `getRotation`) — `camera-controls.ts` / `camera-animation.ts`.

Split along that line into `camera-degenerate-pose-matrix.test.ts` and
`camera-malformed-pose-navigation.test.ts`. Every test body, doc comment
and label moves verbatim; the seam is also a fixture seam, so no shared
helper module is needed — the pose fixtures (`BAD_POSE`, `GOOD_TARGET`,
`poseOf`, `assertPoseUnchanged`, `assertPoseFinite`, `cameraOnBadPose`,
`cameraOnInfinitePose`) were already used by the gesture suite alone,
and only the nine-line `assertAllFinite` predicate is used by both.

Two non-mechanical touches: each file's header now says what it scopes
and points at its sibling, and the "the pose is only half of the
projection matrix" preamble moves down to sit on the describe it
explains rather than on the unrelated frame-clock helper below it.

35 leaf tests before, 35 after, with identical names.
